+++ src/guishortcuts/GUIShortcuts.kt
@@ -95,129 +95,6 @@ infix fun Int.screen(frame: JFrame) {
 }
 
 /**
- * Create a label.
- * Int must be equal to the variable create.
- *
- * @param text the text on the created label.
- *
- * @return a ConstrainedLabel with the supplied text.
- */
-infix fun Int.label(text: String): ConstrainedLabel {
-    if(this != create) throw Exception("Use 'create' when creating components")
-    return ConstrainedLabel(text)
-}
-
-/**
- * Create a button.
- * Int must be equal to the variable create.
- *
- * @param text the text on the created button.
- *
- * @return a ConstrainedButton with the supplied text.
- */
-infix fun Int.button(text: String): ConstrainedButton {
-    if(this != create) throw Exception("Use 'create' when creating components")
-    return ConstrainedButton(text)
-}
-
-/**
- * Create a checkbox.
- * Int must be equal to the variable create.
- *
- * @param text the text on the created checkbox.
- *
- * @return a ConstrainedCheckbox with the supplied text.
- */
-infix fun Int.checkbox(text: String): ConstrainedCheckbox {
-    if(this != create) throw Exception("Use 'create' when creating components")
-    return ConstrainedCheckbox(text)
-}
-
-/**
- * Create a combobox.
- * Int must be equal to the variable create.
- *
- * @param text an array of the values in the combobox.
- *
- * @return a ConstrainedCombobox with the supplied text values.
- */
-infix fun Int.combobox(text: Array<String>): ConstrainedCombobox {
-    if(this != create) throw Exception("Use 'create' when creating components")
-    return ConstrainedCombobox(text)
-}
-
-/**
- * Create a radio button.
- * Int must be equal to the variable create.
- *
- * @param text the text on the created radio button.
- *
- * @return a ConstrainedRadioButton with the supplied text.
- */
-infix fun Int.radiobutton(text: String): ConstrainedRadioButton {
-    if(this != create) throw Exception("Use 'create' when creating components")
-    return ConstrainedRadioButton(text)
-}
-
-/**
- * Create a list.
- * Int must be equal to the variable create.
- *
- * @param text an array of the values in the list.
- *
- * @return a ConstrainedList with the supplied text values.
- */
-infix fun Int.list(text: Array<String>): ConstrainedList {
-    if(this != create) throw Exception("Use 'create' when creating components")
-    return ConstrainedList(text)
-}
-
-/**
- * Create a spinner.
- * Int must be equal to the variable create.
- *
- * @param text an array of the values in the spinner.
- *
- * @return a ConstrainedSpinner with the supplied text values.
- */
-infix fun Int.spinner(text: Array<String>): ConstrainedSpinner {
-    if(this != create) throw Exception("Use 'create' when creating components")
-    return ConstrainedSpinner(text)
-}
-
-/**
- * Create a spinner.
- * Int must be equal to the variable create.
- *
- * @param min the minimum value of the spinner.
- * @param max the maximum value of the spinner.
- *
- * @return a ConstrainedSlider with the supplied minimum and maximum.
- */
-fun slider(min: Int, max: Int): ConstrainedSlider {
-    val slider = ConstrainedSlider(min, max)
-    slider.majorTickSpacing = max - min
-    slider.minorTickSpacing = (max - min) / 2
-    slider.paintTicks = true
-    slider.paintLabels = true
-    return slider
-}
-
-/**
- * Create a text field.
- * Int must be equal to the variable create.
- *
- * @param content the initial content of the text field.
- *
- * @return a ConstrainedTextField with the supplied content.
- */
-infix fun Int.textfield(content: String): ConstrainedTextField {
-    if(this != create) throw Exception("Use 'create' when creating components")
-    val textfield = ConstrainedTextField(content)
-    return textfield
-}
-
-/**
  * Turn a vararg of Strings into an Array<String>.
  *
  * @param options a vararg of the Strings to put in the array.
